fix: Windows large file support and binary-safe downloads
Fixes two Windows (LLP64) issues reported via the VCPKG port:
binary file corruption on download and 32-bit 'long' overflow for files
larger than 2 GiB.

Binary-safe downloads:
- DownloadObject: open the temp output stream with std::ios::binary. On
  Windows the default text mode translates newlines and treats 0x1A as EOF,
  corrupting downloaded binary content (e.g. zip archives).

Large file support (LLP64: 'long' is strictly 32-bit on Windows):
- SelectResult (types.h) and the parser in select.cc: bytes_scanned /
  bytes_processed / bytes_returned switch from 'long int' to 'long long',
  and std::stol to std::stoll
- StatObject (baseclient.cc): parse the content-length header with
  std::stoll so resp.size (size_t) is not truncated above 2 GiB
- ComposeSource (args.h/args.cc): object_size_ member and its assignment
  switch from 'long' to 'long long' so composing sources larger than 2 GiB
  does not overflow

No behavioral change on LP64 platforms (Linux/macOS), where 'long' is
already 64-bit.

Files changed: include/miniocpp/args.h, include/miniocpp/types.h,
src/args.cc, src/baseclient.cc, src/client.cc, src/select.cc

+++ include/miniocpp/args.h
@@ -382,7 +382,7 @@ struct ComposeSource : public ObjectConditionalReadArgs {
   utils::Multimap Headers() const;
 
  private:
-  long object_size_ = -1;
+  long long object_size_ = -1;
   utils::Multimap headers_;
 };  // struct ComposeSource
